requirements.txt
numpy
scikit-learn
sentence-transformers
pytest
chromadb
anthropic
google-generativeai
python-dotenv
